chore: eliminate CSS dependency from chunk worker
