+++ src/neuroglancer/datasource/nggraph/frontend.ts
@@ -21,7 +21,9 @@ import {fetchWithCredentials} from 'neuroglancer/credentials_provider/http_reque
 import {CompleteUrlOptions, CompletionResult, DataSource, DataSourceProvider, DataSubsourceEntry, GetDataSourceOptions} from 'neuroglancer/datasource';
 import {Credentials, NggraphCredentialsProvider} from 'neuroglancer/datasource/nggraph/credentials_provider';
 import {VisibleSegmentsState} from 'neuroglancer/segmentation_display_state/base';
-import {ComputedSplit, isBaseSegmentId, SegmentationGraphSource, SegmentationGraphSourceConnection, UNKNOWN_NEW_SEGMENT_ID, VisibleSegmentEquivalencePolicy} from 'neuroglancer/segmentation_graph/source';
+import {isBaseSegmentId, UNKNOWN_NEW_SEGMENT_ID, VisibleSegmentEquivalencePolicy} from 'neuroglancer/segmentation_graph/segment_id';
+import {ComputedSplit, SegmentationGraphSource, SegmentationGraphSourceConnection} from 'neuroglancer/segmentation_graph/source';
+import {SegmentationUserLayer} from 'neuroglancer/segmentation_user_layer';
 import {StatusMessage} from 'neuroglancer/status';
 import {Uint64Set} from 'neuroglancer/uint64_set';
 import {CancellationToken, uncancelableToken} from 'neuroglancer/util/cancellation';
@@ -30,7 +32,6 @@ import {DisjointUint64Sets} from 'neuroglancer/util/disjoint_sets';
 import {responseJson} from 'neuroglancer/util/http_request';
 import {parseArray, verifyFiniteFloat, verifyInt, verifyObject, verifyObjectProperty, verifyString, verifyStringArray} from 'neuroglancer/util/json';
 import {Uint64} from 'neuroglancer/util/uint64';
-import {SegmentationUserLayer} from 'neuroglancer/segmentation_user_layer';
 
 const urlPattern = '^(https?://[^/]+)/(.*)$';
